1：异常处理

+++ src/test_lambda.rs
@@ -1,4 +1,4 @@
-pub fn exec_lambda(a: u32, b: u32, c: u32) -> u32 {
+pub fn exec_lambda(a: f64, b: f64, c: f64) -> f64 {
     let calculate = |a, b, c| a * b + c;
     let mut idx:i64 = 0;
     while idx < 100 {
